backend/requirements.txt
fastapi
uvicorn
pydantic
networkx
numpy>=2.1.0
google-generativeai
python-dotenv
openai

